add config option for images

+++ twitter_utils.py
@@ -18,25 +18,28 @@ api = tweepy.API(auth)
 def get_tweets(user="pokevisiongo"):
 	return api.user_timeline(screen_name=user, count=20)
 
-def analyze_tweets(tweets):
+def analyze_tweets(tweets, pictures=False):
 	analysis = {}
 	for tweet in tweets:
 		try:
 			analysis[tweet.id] = analyze(tweet.text)
-			media_files = tweet.entities.get("media", [])
-			for file in media_files:
-				filename = wget.download(file["media_url"])
-				if is_nude(str(filename)):
-					analysis[tweet.id] = "bad image"
-					break
-			os.system('rm *.jpg *.png')
+			if pictures:
+				media_files = tweet.entities.get("media", [])
+				for file in media_files:
+					filename = wget.download(file["media_url"])
+					if is_nude(str(filename)):
+						analysis[tweet.id] = "bad image"
+						break
+				os.system('rm *.jpg *.png')
 		except:
-			media_files = tweet.entities.get("media", [])
 			analysis[tweet.id] = "none"
-			for file in media_files:
-				filename = wget.download(file["media_url"])
-				if is_nude(str(filename)):
-					analysis[tweet.id] = "bad image"
-					break
-			os.system('rm *.jpg')
+			if pictures:
+				media_files = tweet.entities.get("media", [])
+				for file in media_files:
+					filename = wget.download(file["media_url"])
+					if is_nude(str(filename)):
+						analysis[tweet.id] = "bad image"
+						break
+				os.system('rm *.jpg')
+
 	return analysis
